+++ src/__tests__/index.spec.ts
@@ -42,6 +42,7 @@ test('setSuperSourceProperties - 7.2', async () => {
 		expect(conn.sendCommand).toHaveBeenCalledTimes(1)
 		expect(conn.sendCommand).toHaveBeenNthCalledWith(1, {
 			flag: 12,
+			runOrderGroup: 0,
 			_properties: {
 				artOption: 0,
 				artPreMultiplied: true,
@@ -68,6 +69,7 @@ test('setSuperSourceProperties - 8.0', async () => {
 		expect(conn.sendCommand).toHaveBeenNthCalledWith(1, {
 			ssrcId: 2,
 			flag: 12,
+			runOrderGroup: 0,
 			_properties: {
 				artOption: 0,
 				artPreMultiplied: true,
@@ -93,6 +95,7 @@ test('setSuperSourceBorder - 7.2', async () => {
 		expect(conn.sendCommand).toHaveBeenCalledTimes(1)
 		expect(conn.sendCommand).toHaveBeenNthCalledWith(1, {
 			flag: 139264,
+			runOrderGroup: 0,
 			_properties: {
 				borderBevelSoftness: 12,
 				borderLuma: 3,
@@ -119,6 +122,7 @@ test('setSuperSourceBorder - 8.0', async () => {
 		expect(conn.sendCommand).toHaveBeenNthCalledWith(1, {
 			ssrcId: 2,
 			flag: 1088,
+			runOrderGroup: 0,
 			_properties: {
 				borderBevelSoftness: 12,
 				borderLuma: 3,