+++ src/commands/SuperSource/index.ts
@@ -1,2 +1,3 @@
+export * from './SuperSourceBoxBorderCommand'
 export * from './SuperSourceBoxParametersCommand'
 export * from './SuperSourcePropertiesCommand'